feat(rush-lib): rush-pnpm accepts patch & patch-commit
